Feat : useModal 구현, Modal style 분리
